+++ src/Data/Vector/Unboxed/Bit.hs
@@ -32,15 +32,9 @@ module Data.Vector.Unboxed.Bit
      , and
      , or
 
-     , any
-     , anyBits
-     , all
-     , allBits
-
      , reverse
 
      , first
-     , findIndex
      ) where
 
 import           Control.Monad
@@ -198,27 +192,6 @@ or v = loop 0
             | otherwise = (indexWord v i /= 0)
                         || loop (i + wordSize)
 
-all :: (Bit -> Bool) -> Vector Bit -> Bool
-all p = case (p (Bit False), p (Bit True)) of
-    (False, False) -> U.null
-    (False,  True) -> allBits (Bit True)
-    (True,  False) -> allBits (Bit False)
-    (True,   True) -> flip seq True
-
-any :: (Bit -> Bool) -> Vector Bit -> Bool
-any p = case (p (Bit False), p (Bit True)) of
-    (False, False) -> flip seq False
-    (False,  True) -> anyBits (Bit True)
-    (True,  False) -> anyBits (Bit False)
-    (True,   True) -> not . U.null
-
-allBits, anyBits :: Bit -> U.Vector Bit -> Bool
-allBits (Bit False) = not . or
-allBits (Bit True) = and
-
-anyBits (Bit False) = not . and
-anyBits (Bit True) = or
-
 reverse :: U.Vector Bit -> U.Vector Bit
 reverse xs = runST $ do
     let !n = V.length xs
@@ -238,10 +211,3 @@ first b xs = mfilter (< n) (loop 0)
         loop !i
             | i >= n    = Nothing
             | otherwise = fmap (i +) (ff (indexWord xs i)) `mplus` loop (i + wordSize)
-
-findIndex :: (Bit -> Bool) -> Vector Bit -> Maybe Int
-findIndex p xs = case (p (Bit False), p (Bit True)) of
-    (False, False) -> Nothing
-    (False,  True) -> first (Bit True)  xs
-    (True,  False) -> first (Bit False) xs
-    (True,   True) -> if V.null xs then Nothing else Just 0