+++ src/Data/Vector/Unboxed/Mutable/Bit.hs
@@ -30,11 +30,6 @@ module Data.Vector.Unboxed.Mutable.Bit
      , and
      , or
 
-     , any
-     , anyBits
-     , all
-     , allBits
-
      , reverseInPlace
      ) where
 
@@ -222,27 +217,6 @@ or v = loop 0
                 if y /= 0
                     then return True
                     else loop (i + wordSize)
-
-all :: PrimMonad m => (Bit -> Bool) -> U.MVector (PrimState m) Bit -> m Bool
-all p = case (p (Bit False), p (Bit True)) of
-    (False, False) -> return . MV.null
-    (False,  True) -> allBits (Bit True)
-    (True,  False) -> allBits (Bit False)
-    (True,   True) -> flip seq (return True)
-
-any :: PrimMonad m => (Bit -> Bool) -> U.MVector (PrimState m) Bit -> m Bool
-any p = case (p (Bit False), p (Bit True)) of
-    (False, False) -> flip seq (return False)
-    (False,  True) -> anyBits (Bit True)
-    (True,  False) -> anyBits (Bit False)
-    (True,   True) -> return . not . MV.null
-
-allBits, anyBits :: PrimMonad m => Bit -> U.MVector (PrimState m) Bit -> m Bool
-allBits (Bit False) = liftM not . or
-allBits (Bit True) = and
-
-anyBits (Bit False) = liftM not . and
-anyBits (Bit True) = or
 
 reverseInPlace :: PrimMonad m => U.MVector (PrimState m) Bit -> m ()
 reverseInPlace xs = loop 0 (MV.length xs)