+++ changelog.md
@@ -3,6 +3,7 @@
 * Change arguments of 'unions' and 'intersections'.
 * 'fromWords' and 'cloneFromWords' no longer take a number of bits.
 * Remove 'pad' and 'padWith'.
+* Remove 'any', 'anyBits', 'all', 'allBits', 'findIndex'.
 
 # 0.2.0.1